+++ Tests/test_integration.py
@@ -1,18 +1,20 @@
 from __future__ import print_function
 import copy
 import time
+import re
+from subprocess import Popen, PIPE
 from pprint import pformat
 import uuid
 import ast
-import urllib
+import urllib.parse
 import urllib3
 import requests.exceptions
 from demisto_client.demisto_api.rest import ApiException
 import demisto_client
 import json
 
-from Tests.test_utils import print_error, print_warning, print_color, LOG_COLORS, Docker
-from Tests.scripts.constants import PB_Status
+from demisto_sdk.commands.common.tools import print_error, print_warning, print_color, LOG_COLORS
+from demisto_sdk.commands.common.constants import PB_Status
 
 # Disable insecure warnings
 urllib3.disable_warnings()
@@ -21,6 +23,311 @@ urllib3.disable_warnings()
 DEFAULT_TIMEOUT = 60
 DEFAULT_INTERVAL = 20
 ENTRY_TYPE_ERROR = 4
+
+
+# ----- Docker class ----- #
+class Docker:
+    """ Client for running docker commands on remote machine using ssh connection.
+
+    """
+    PYTHON_INTEGRATION_TYPE = 'python'
+    JAVASCRIPT_INTEGRATION_TYPE = 'javascript'
+    DEFAULT_PYTHON2_IMAGE = 'demisto/python'
+    DEFAULT_PYTHON3_IMAGE = 'demisto/python3'
+    COMMAND_FORMAT = '{{json .}}'
+    MEMORY_USAGE = 'MemUsage'
+    PIDS_USAGE = 'PIDs'
+    CONTAINER_NAME = 'Name'
+    CONTAINER_ID = 'ID'
+    DEFAULT_CONTAINER_MEMORY_USAGE = 75
+    DEFAULT_CONTAINER_PIDS_USAGE = 3
+    REMOTE_MACHINE_USER = 'ec2-user'
+    SSH_OPTIONS = 'ssh -o StrictHostKeyChecking=no'
+
+    @classmethod
+    def _build_ssh_command(cls, server_ip, remote_command, force_tty=False):
+        """Add and returns ssh prefix and escapes remote command
+
+            Args:
+                server_ip (str): remote machine ip to connect using ssh.
+                remote_command (str): command to execute in remote machine.
+                force_tty (bool): adds -t flag in order to force tty allocation.
+
+            Returns:
+                str: full ssh command
+
+        """
+        remote_server = '{}@{}'.format(cls.REMOTE_MACHINE_USER, server_ip)
+        ssh_prefix = '{} {}'.format(cls.SSH_OPTIONS, remote_server)
+        if force_tty:
+            ssh_prefix += ' -t'
+        # escaping the remote command with single quotes
+        cmd = "{} '{}'".format(ssh_prefix, remote_command)
+
+        return cmd
+
+    @classmethod
+    def _build_stats_cmd(cls, server_ip, docker_images):
+        """ Builds docker stats and grep command string.
+
+        Example of returned value:
+        ssh -o StrictHostKeyChecking=no ec2-user@server_ip
+        'sudo docker stats --no-stream --no-trunc --format "{{json .}}" | grep -Ei "demistopython33.7.2.214--"'
+        Grep is based on docker images names regex.
+
+            Args:
+                server_ip (str): Remote machine ip to connect using ssh.
+                docker_images (set): Set of docker images.
+
+            Returns:
+                str: String command to run later as subprocess.
+
+        """
+        # docker stats command with json output
+        docker_command = 'sudo docker stats --no-stream --no-trunc --format "{}"'.format(cls.COMMAND_FORMAT)
+        # replacing : and / in docker images names in order to grep the stats by container name
+        docker_images_regex = ['{}--'.format(re.sub('[:/]', '', docker_image)) for docker_image in docker_images]
+        pipe = ' | '
+        grep_command = 'grep -Ei "{}"'.format('|'.join(docker_images_regex))
+        remote_command = docker_command + pipe + grep_command
+        cmd = cls._build_ssh_command(server_ip, remote_command)
+
+        return cmd
+
+    @classmethod
+    def _build_kill_cmd(cls, server_ip, container_name):
+        """ Constructs docker kll command string to run on remote machine.
+
+            Args:
+                server_ip (str): Remote machine ip to connect using ssh.
+                container_name (str): Docker container name to kill.
+
+            Returns:
+                str: String of docker kill command on remote machine.
+        """
+        remote_command = 'sudo docker kill {}'.format(container_name)
+        cmd = cls._build_ssh_command(server_ip, remote_command)
+
+        return cmd
+
+    @classmethod
+    def _build_pid_info_cmd(cls, server_ip, container_id):
+        """Constructs docker exec ps command string to run on remote machine.
+
+            Args:
+                server_ip (str): Remote machine ip to connect using ssh.
+                container_id (str): Docker container id.
+
+            Returns:
+                str: String of docker exec ps command on remote machine.
+
+        """
+        remote_command = 'sudo docker exec -it {} ps -fe'.format(container_id)
+        cmd = cls._build_ssh_command(server_ip, remote_command, force_tty=True)
+
+        return cmd
+
+    @classmethod
+    def _parse_stats_result(cls, stats_lines):
+        """Parses the docker statics str and converts to Mib.
+
+            Args:
+                stats_lines (str): String that contains docker stats.
+            Returns:
+                list: List of dictionaries with parsed docker container statistics.
+
+        """
+        stats_result = []
+        try:
+            containers_stats = [json.loads(c) for c in stats_lines.splitlines()]
+
+            for container_stat in containers_stats:
+                memory_usage_stats = container_stat.get(cls.MEMORY_USAGE, '').split('/')[0].lower()
+
+                if 'kib' in memory_usage_stats:
+                    mib_usage = float(memory_usage_stats.replace('kib', '').strip()) / 1024
+                elif 'gib' in memory_usage_stats:
+                    mib_usage = float(memory_usage_stats.replace('kib', '').strip()) * 1024
+                else:
+                    mib_usage = float(memory_usage_stats.replace('mib', '').strip())
+
+                stats_result.append({
+                    'memory_usage': mib_usage,
+                    'pids': int(container_stat.get(cls.PIDS_USAGE)),
+                    'container_name': container_stat.get(cls.CONTAINER_NAME),
+                    'container_id': container_stat.get(cls.CONTAINER_ID)
+                })
+        except Exception as e:
+            print_warning("Failed in parsing docker stats result, returned empty list. Additional info: {}".format(e))
+        finally:
+            return stats_result
+
+    @classmethod
+    def run_shell_command(cls, cmd):
+        """Executes shell command and returns outputs of the process.
+
+            Args:
+                cmd (str): command to execute.
+
+            Returns:
+                str: stdout of the executed command.
+                str: stderr of the executed command.
+
+        """
+        process = Popen(cmd, stdout=PIPE, stderr=PIPE, shell=True, universal_newlines=True)
+        stdout, stderr = process.communicate()
+
+        return stdout, stderr
+
+    @classmethod
+    def get_image_for_container_id(cls, server_ip, container_id):
+        cmd = cls._build_ssh_command(server_ip, "sudo docker inspect -f {{.Config.Image}} " + container_id,
+                                     force_tty=False)
+        stdout, stderr = cls.run_shell_command(cmd)
+        if stderr:
+            print_warning("Received stderr from docker inspect command. Additional information: {}".format(stderr))
+        res = stdout or ""
+        return res.strip()
+
+    @classmethod
+    def get_integration_image(cls, integration_config):
+        """ Returns docker image of integration that was configured using rest api call via demisto_client
+
+            Args:
+                integration_config (dict): Integration config that included script section.
+            Returns:
+                list: List that includes integration docker image name. If no docker image was found,
+                      default python2 and python3 images are returned.
+
+        """
+        integration_script = integration_config.get('configuration', {}).get('integrationScript', {}) or {}
+        integration_type = integration_script.get('type')
+        docker_image = integration_script.get('dockerImage')
+
+        if integration_type == cls.JAVASCRIPT_INTEGRATION_TYPE:
+            return None
+        elif integration_type == cls.PYTHON_INTEGRATION_TYPE and docker_image:
+            return [docker_image]
+        else:
+            return [cls.DEFAULT_PYTHON2_IMAGE, cls.DEFAULT_PYTHON3_IMAGE]
+
+    @classmethod
+    def docker_stats(cls, server_ip, docker_images):
+        """ Executes docker stats command and greps all containers with prefix of docker images names.
+
+            Args:
+                server_ip (str): Remote machine ip to connect using ssh.
+                docker_images (set): Set of docker images to check their resource usage.
+
+            Returns:
+                list: List of dictionaries with parsed container memory statistics.
+        """
+        cmd = cls._build_stats_cmd(server_ip, docker_images)
+        stdout, stderr = cls.run_shell_command(cmd)
+
+        if stderr:
+            print_warning("Failed running docker stats command. Additional information: {}".format(stderr))
+            return []
+
+        return cls._parse_stats_result(stdout)
+
+    @classmethod
+    def kill_container(cls, server_ip, container_name):
+        """ Executes docker kill command on remote machine using ssh.
+
+            Args:
+                server_ip (str): The remote server ip address.
+                container_name (str): The container name to kill
+
+        """
+        cmd = cls._build_kill_cmd(server_ip, container_name)
+        _, stderr = cls.run_shell_command(cmd)
+
+        if stderr:
+            print_warning("Failed killing container: {}\nAdditional information: {}".format(container_name, stderr))
+
+    @classmethod
+    def get_docker_pid_info(cls, server_ip, container_id):
+        """Executes docker exec ps command on remote machine using ssh.
+
+            Args:
+                server_ip (str): The remote server ip address.
+                container_id (str): Docker container id.
+
+            Returns:
+                str: output of executed command.
+        """
+        cmd = cls._build_pid_info_cmd(server_ip, container_id)
+        stdout, stderr = cls.run_shell_command(cmd)
+
+        if stderr:
+            ignored_warning_message = "Connection to {} closed".format(server_ip)
+            if ignored_warning_message not in stderr:
+                print_warning("Failed getting pid info for container id: {}.\nAdditional information: {}".
+                              format(container_id, stderr))
+
+        return stdout
+
+    @classmethod
+    def check_resource_usage(cls, server_url, docker_images, def_memory_threshold, def_pid_threshold,
+                             docker_thresholds):
+        """
+        Executes docker stats command on remote machine and returns error message in case of exceeding threshold.
+
+        Args:
+            server_url (str): Target machine full url.
+            docker_images (set): Set of docker images to check their resource usage.
+            def_memory_threshold (int): Memory threshold of specific docker container, in Mib.
+            def_pids_threshold (int): PIDs threshold of specific docker container, in Mib.
+            docker_thresholds: thresholds per docker image
+
+        Returns:
+            str: The error message. Empty in case that resource check passed.
+
+        """
+        server_ip = server_url.lstrip("https://")
+        containers_stats = cls.docker_stats(server_ip, docker_images)
+        error_message = ""
+
+        for container_stat in containers_stats:
+            failed_memory_test = False
+            container_name = container_stat['container_name']
+            container_id = container_stat['container_id']
+            memory_usage = container_stat['memory_usage']
+            pids_usage = container_stat['pids']
+            image_full = cls.get_image_for_container_id(server_ip,
+                                                        container_id)  # get full name (ex: demisto/slack:1.0.0.4978)
+            image_name = image_full.split(':')[0]  # just the name such as demisto/slack
+
+            memory_threshold = (docker_thresholds.get(image_full, {}).get('memory_threshold') or docker_thresholds.get(
+                image_name, {}).get('memory_threshold') or def_memory_threshold)
+            pid_threshold = (docker_thresholds.get(image_full, {}).get('pid_threshold')
+                             or docker_thresholds.get(image_name, {}).get('pid_threshold') or def_pid_threshold)
+            print("Checking container: {} (image: {}) for memory: {} pid: {} thresholds ...".format(
+                container_name, image_full, memory_threshold, pid_threshold))
+            if memory_usage > memory_threshold:
+                error_message += ('Failed docker resource test. Docker container {} exceeded the memory threshold, '
+                                  'configured: {} MiB and actual memory usage is {} MiB.\n'
+                                  'Fix container memory usage or add `memory_threshold` key to failed test '
+                                  'in conf.json with value that is greater than {}\n'
+                                  .format(container_name, memory_threshold, memory_usage, memory_usage))
+                failed_memory_test = True
+            if pids_usage > pid_threshold:
+                error_message += ('Failed docker resource test. Docker container {} exceeded the pids threshold, '
+                                  'configured: {} and actual pid number is {}.\n'
+                                  'Fix container pid usage or add `pid_threshold` key to failed test '
+                                  'in conf.json with value that is greater than {}\n'
+                                  .format(container_name, pid_threshold, pids_usage, pids_usage))
+                additional_pid_info = cls.get_docker_pid_info(server_ip, container_id)
+                if additional_pid_info:
+                    error_message += 'Additional pid information:\n{}'.format(additional_pid_info)
+                failed_memory_test = True
+
+            if failed_memory_test:
+                # killing current container in case of memory resource test failure
+                cls.kill_container(server_ip, container_name)
+
+        return error_message
 
 
 # ----- Functions ----- #
@@ -212,6 +519,7 @@ def __disable_integrations_instances(client, module_instances, prints_manager, t
             error_message = 'Failed to disable integration instance, error trying to communicate with demisto ' \
                             'server: {} '.format(conn_err)
             prints_manager.add_print_job(error_message, print_error, thread_index)
+            return
 
         if res[1] != 200:
             error_message = 'disable instance failed with status code ' + str(res[1])
@@ -349,7 +657,7 @@ def __delete_incident(client, incident, prints_manager, thread_index=0):
 def __delete_integration_instance(client, instance_id, prints_manager, thread_index=0):
     try:
         res = demisto_client.generic_request_func(self=client, method='DELETE',
-                                                  path='/settings/integration/' + urllib.quote(
+                                                  path='/settings/integration/' + urllib.parse.quote(
                                                       instance_id))
     except requests.exceptions.RequestException as conn_err:
         error_message = 'Failed to delete integration instance, error trying to communicate with demisto ' \
@@ -378,7 +686,7 @@ def __print_investigation_error(client, playbook_id, investigation_id, prints_ma
     try:
         empty_json = {"pageSize": 1000}
         res = demisto_client.generic_request_func(self=client, method='POST',
-                                                  path='/investigation/' + urllib.quote(
+                                                  path='/investigation/' + urllib.parse.quote(
                                                       investigation_id), body=empty_json)
     except requests.exceptions.RequestException as conn_err:
         error_message = 'Failed to print investigation error, error trying to communicate with demisto ' \
@@ -387,15 +695,15 @@ def __print_investigation_error(client, playbook_id, investigation_id, prints_ma
     if res and int(res[1]) == 200:
         resp_json = ast.literal_eval(res[0])
         entries = resp_json['entries']
-        prints_manager.add_print_job('Playbook ' + playbook_id + ' has failed:', print_color, thread_index,
+        prints_manager.add_print_job('Playbook {} has failed:'.format(playbook_id), print_color, thread_index,
                                      message_color=color)
         for entry in entries:
             if entry['type'] == ENTRY_TYPE_ERROR and entry['parentContent']:
-                prints_manager.add_print_job('- Task ID: ' + entry['taskId'].encode('utf-8'), print_color, thread_index,
+                prints_manager.add_print_job('- Task ID: {}'.format(entry['taskId']), print_color, thread_index,
                                              message_color=color)
-                prints_manager.add_print_job('  Command: ' + entry['parentContent'].encode('utf-8'), print_color,
+                prints_manager.add_print_job('  Command: {}'.format(entry['parentContent']), print_color,
                                              thread_index, message_color=color)
-                body_contents_str = '  Body:\n' + entry['contents'].encode('utf-8') + '\n'
+                body_contents_str = '  Body:\n{}\n'.format(entry['contents'])
                 prints_manager.add_print_job(body_contents_str, print_color,
                                              thread_index, message_color=color)