+++ Packs/Base/Scripts/CommonServerPython/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-Added a default value to the indicator timeline field **Category** when a value is not provided in an entry's timeline data.
+  - Added the argument *ignore_auto_extract* to the ***return_outputs*** command.
+  - Added a default value to the indicator timeline field **Category** when a value is not provided in an entry's timeline data.
 
 ## [20.3.4] - 2020-03-30
 - Added support for successful empty responses (status code 204) in the base client.